service2/requirements.txt
strawberry-graphql[fastapi]==0.96.0
uvicorn==0.15.0
langchain==0.0.267
langchain-openai==0.0.2.post1
opensearch-py==2.3.1
python-multipart==0.0.5
pypdf==3.15.4
aiohttp==3.8.5
httpx==0.24.1
boto3==1.26.137